more relaxed string definition for uint8 vectors
Most  recent  VoLTE  items  provide  often  large  byte  fields  which
represent string values. To simplify  the definition of these the type
checking is more relaxed. Be aware that in case of doubt e.g. when the
string could  also be interpreted as  a value list, there  is still an
error generated.

Signed-off-by: Otto Linnemann <otto.linnemann@peiker.de>