support for EFS_Backup store
NV items for radio frequency  calibration beginning from 20000 have to
be  stored in  the dedicated  stream EFS_Backup  within OLE2  compound
document where the qcn data format is based on. The section indicators
within the ASCII print out are adjusted as well.

This  is  work in  progress.  The  RF  NV  items (>=20000)  are  still
not  fully compatible  to the  original Qualcomm  implementation since
compression is not implemented yet.

Signed-off-by: Otto Linnemann <otto.linnemann@peiker.de>