support compressed attribute in item schema
The  schema definition  for efs  and rf  items allows  to specify  the
attribute 'compressed' to instruct QMSL to compress the data with ZLIB
algorithm before its  written to the QCN file. This  feature is mostly
used for  the radio frequency  (rf) items  and was not  required until
now. It is implemented by the present commit.

Signed-off-by: Otto Linnemann <otto.linnemann@peiker.de>