flat output of all EFS items for better comparison
For reasons which I do really not fully understand the so called rf nv
items with ID's 20000 or above are not stored consistently to same EFS
area as stated  whithin the comment in the schema  file. QMDL seems to
put most of them to the qcn directory 'EFS_Backup' but some e.g. 23387
are  stored  to the  EFS  directory  'NV_Items' instead.  Nevertheless
Qualcomm uitilies like  QRCT, QMDL, QPST seem to be  able to deal with
both location so this seems to not be a problem first of all.

Anyway, by ending  up with different QCN directories of  QPST and ORCT
compiled files we are better adviced to  list all EFS based items in a
flat list  for the sake  of simpler comparison. This  commit therefore
introduces a so called flat  output for all print-item functions which
is enabled by default but can  be disabled by increasing the verbosity
level.

Signed-off-by: Otto Linnemann <otto.linnemann@peiker.de>