support compilation of multiple files
Instead of  invoking the  uberjar application  to compile  each source
file separately  it is much more  efficient to let clojure  do the job
just to avoid  to startup the JVM  multiple times and to  parse the nv
definition file again and again.  The latter parsing operation takes a
considerable amount of time.

Signed-off-by: Otto Linnemann <linneman@gmx.de>